refactor: feature_extractor moved to viscy.representation.evaluation
